requirements.txt
flask
flask-cors
joblib
numpy
torch
sentence-transformers
python-docx
PyMuPDF
scikit-learn
transformers
safetensors